refactor(journeys-admin): rename CardTemplates inner dir Templates/ to Layouts/ (ENG-3725 A21)
The inner dir holds card layout presets (CardVideo, CardIntro, ...);
'Templates' collided with the journey Templates domain term.
