+++ apps/journeys-admin/src/components/Editor/Slider/Settings/Drawer/CardTemplates/Layouts/CardForm/CardForm.spec.tsx
