+++ apps/journeys-admin/src/components/Editor/Slider/Settings/Drawer/CardTemplates/Layouts/CardPoll/index.ts
