+++ apps/journeys-admin/src/components/Editor/Slider/Settings/Drawer/CardTemplates/Layouts/CardQuote/index.ts
